add task item in module
